Use fseeko64/ftello64 on MinGW instead of broken _fseeki64/_ftelli64.
At least _ftelli64 is broken, sometimes returning position 3 after reading 4 bytes from the beginning of a stream with fread. It disagrees with the not-broken ftell in this case.

Issue was triggered by adding FLAC files to the fre:ac joblist. Sometimes reading a FLAC would fail which was traced back to this _ftelli64 issue.

Might reconsider using _ftelli64 after updating the toolchain.